Reorder rule customization
Put `additionalRules` before `alterRules` and `filterRules` to ensure altering
and filtering also works on the contents of `additionalRules`.

Note that the use-case for `alterRules` on `additionalRules` is quite
flimsy, as any alterations need to be made directly in the Jsonnet
source anyway, so alterations should probably be made directly in the additional
rule definitions.

On the other hand, filtering `additionalRules` could be valuable, as the
rule filtering can be customized via hierarchy.