Make alert namespace selector configurable
This is necessary because the changes introduced in #19 will make the
namespace selector actually work for some alerts, e.g.
`KubePersistentVolumeFillingUp`.

With this change, users of the component can choose to receive alerts
for a different set of namespaces than the default selector, which
includes only namespaces which are part of Kubernetes, Rancher, or
Project Syn.