Adjust Rancher federation to keep `namespace` label if it is correct
We implemented relabeling configs to ensure all federated metrics have
the correct `namespace` label in #19. However, the relabeling config
provided in that PR drops the original `namespace` label for metrics
which don't need fixing.

This commit adds an extra relabeling step to make sure the original
namespace label is put back, if it was correct.

Fixes #26